Refactor: Move blazepromote and blazeoverlay to blaze package. Remove nested ui package
